requirements.txt
streamlit
google-generativeai
gTTS
streamlit-mic-recorder
python-dotenv